Feature: Implemented the delete forever from archive
